Added PeerProcess
PeerProcess is the main java class which unites server and client behaviors. Each instance of PeerProcess represents a machine in the P2P network. It threads connections to communicate with multiple peers in a psuedo-parallel way.
Need to implement thread safe information for the peers for later use with unchoking/choking and how to stop threads when timers are up.